pyproject.toml
[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bitescore"
version = "0.2.0"
description = "Predict and rank digestibility of protein sequences with an ML-powered recommender"
readme = "README.md"
authors = [{name="Bitescore Devs"}]
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "biopython>=1.83",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "requests>=2.32",
    "click>=8.1",
    "gradio>=4.38",
]

[project.optional-dependencies]
go = ["goatools>=1.3.1"]
dev = ["pytest>=7.4", "pytest-cov>=4.1", "ruff>=0.6.8"]

[project.scripts]
bitescore = "bitescore.cli:main"
bitescore-chat = "bitescore.app.chatbot:main"

[tool.setuptools]
package-dir = {"bitescore" = "src/bitescore"}

[tool.setuptools.package-data]
bitescore = ["data/*"]
